requirements.txt
Flask==3.0.0
requests==2.31.0
PyYAML==6.0.1
APScheduler==3.10.4
pandas==2.1.4

